fix: address v0.8.1 review findings
- revert raw_source truncation in ParsingError (GC handles this)
- move parse_money_value_string back to parsers/utils, remove types/utils.py
- use local import in order_page.order_total to break circular import
