+++ funpayparsers/parsers/utils.py
@@ -19,11 +19,12 @@ from collections.abc import Iterable
 from selectolax.lexbor import LexborNode, LexborHTMLParser
 
 from funpayparsers.types.enums import BadgeType
-from funpayparsers.types.utils import parse_money_value_string
+from funpayparsers.types.common import MoneyValue
 from funpayparsers.types.messages import Message
 
 
 CSS_URL_RE = re.compile(r'url\(([^()]+)\)', re.IGNORECASE)
+MONEY_VALUE_RE = re.compile(r'^([+\-]?\d+(?:\.\d+)?)(.+)$')
 
 
 TODAY_WORDS = ['сегодня', 'сьогодні', 'today']
@@ -224,6 +225,52 @@ def resolve_messages_senders(messages: Iterable[Message], /) -> None:
             badge,
             send_time,
         )
+
+
+@overload
+def parse_money_value_string(
+    money_value_str: str, /, *, raw_source: str | None = ..., raise_on_error: Literal[True] = ...
+) -> MoneyValue: ...
+
+
+@overload
+def parse_money_value_string(
+    money_value_str: str, /, *, raw_source: str | None = ..., raise_on_error: Literal[False] = ...
+) -> MoneyValue | None: ...
+
+
+def parse_money_value_string(
+    money_value_str: str,
+    /,
+    *,
+    raw_source: str | None = None,
+    raise_on_error: bool = False,
+) -> MoneyValue | None:
+    """
+    Parse money value string.
+
+    Possible formats:
+        - `+ 1.23 ₽`
+        - `- 1.23 $`
+        - `1.23 €`
+        - `etc.`
+
+    Whitespaces between sign, value and currency char are allowed.
+    String will be stripped before parsing.
+    """
+    to_process = money_value_str.strip().replace(' ', '').replace('−', '-')
+    if not (match := MONEY_VALUE_RE.fullmatch(to_process)):
+        if raise_on_error:
+            raise ValueError(f"Unable to parse money value string '{money_value_str}'")
+        return None
+
+    value, currency = match.groups()
+
+    return MoneyValue(
+        raw_source=raw_source if raw_source is not None else money_value_str,
+        value=float(value),
+        character=currency,
+    )
 
 
 def serialize_form(source: str | LexborNode | LexborHTMLParser) -> dict[str, str]:
